Implementacion limpia hu1
se mejoro la implementacion de hu1 para que sea mas limpia y se ejecuten todos los checks en usecases de category

+++ src/main/java/com/example/Emazon/domain/usecase/CategoryUseCases.java
@@ -1,9 +1,7 @@
-//domain.usecase.CategoryUseCases
 package com.example.emazon.domain.usecase;
 
 import com.example.emazon.domain.api.ICategoryServicePort;
-import com.example.emazon.domain.exceptions.CategoryAlreadyExistsException;
-import com.example.emazon.domain.exceptions.NoDataFoundException;
+import com.example.emazon.domain.exceptions.*;
 import com.example.emazon.domain.model.Category;
 import com.example.emazon.domain.spi.ICategoryPersistencePort;
 
@@ -20,11 +18,14 @@ public class CategoryUseCases implements ICategoryServicePort {
 
     @Override
     public void saveCategory(Category category) {
+        validateCategory(category);
+
         // Validación de categoría duplicada
         Optional<Category> existingCategory = categoryPersistencePort.findByName(category.getName());
         if (existingCategory.isPresent()) {
             throw new CategoryAlreadyExistsException();
         }
+
         categoryPersistencePort.saveCategory(category);
     }
 
@@ -37,16 +38,35 @@ public class CategoryUseCases implements ICategoryServicePort {
     public Category getCategory(Long id) {
         // Validación de existencia de categoría
         return categoryPersistencePort.getCategory(id)
-                .orElseThrow(NoDataFoundException::new);
+                .orElseThrow(CategoryNotFoundException::new);
     }
 
     @Override
     public void updateCategory(Category category) {
+        validateCategory(category);
         categoryPersistencePort.updateCategory(category);
     }
 
     @Override
     public void deleteCategory(Long id) {
+        if (!categoryPersistencePort.getCategory(id).isPresent()) {
+            throw new CategoryNotFoundException();
+        }
         categoryPersistencePort.deleteCategory(id);
+    }
+
+    private void validateCategory(Category category) {
+        if (category.getName() == null || category.getName().isEmpty()) {
+            throw new InvalidCategoryNameException();
+        }
+        if (category.getDescription() == null || category.getDescription().isEmpty()) {
+            throw new InvalidCategoryDescriptionException();
+        }
+        if (category.getName().length() > 50) {
+            throw new CategoryNameTooLongException();
+        }
+        if (category.getDescription().length() > 90) {
+            throw new CategoryDescriptionTooLongException();
+        }
     }
 }
